feat: move txOptions to intefaces
